+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add CLI alias configuration to `nxos_system` resource and data source
 - Add storm control policy attributes to `nxos_physical_interface` resource and data source
 - Add storm control policy attributes to `nxos_port_channel_interface` resource and data source
 - Add boot configuration and boot image attributes to `nxos_system` resource and data source